refactor: decompose MCPOAuth
